requirements.txt
altgraph==0.17.4
macholib==1.16.3
Markdown==3.7
packaging==24.2
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
PyYAML==6.0.2
schema==0.7.7
setuptools==78.0.2
